@@ -9,11 +9,6 @@
   "bugs": {
     "url": "https://github.com/akveo/ngx-admin/issues"
   },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "IE 11"
-  ],
   "scripts": {
     "ng": "ng",
     "conventional-changelog": "conventional-changelog",
@@ -35,27 +30,26 @@
   },
   "dependencies": {
     "@agm/core": "^1.0.0-beta.5",
-    "@angular/animations": "^7.2.11",
-    "@angular/cdk": "^7.3.6",
-    "@angular/common": "^7.2.11",
-    "@angular/compiler": "^7.2.11",
-    "@angular/core": "^7.2.11",
-    "@angular/forms": "^7.2.11",
-    "@angular/http": "^7.2.11",
-    "@angular/platform-browser": "^7.2.11",
-    "@angular/platform-browser-dynamic": "^7.2.11",
-    "@angular/router": "^7.2.11",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "@asymmetrik/ngx-leaflet": "3.0.1",
-    "@nebular/auth": "3.4.2",
-    "@nebular/bootstrap": "3.4.2",
-    "@nebular/security": "3.4.2",
-    "@nebular/theme": "3.4.2",
-    "@ng-bootstrap/ng-bootstrap": "^4.0.0",
+    "@nebular/auth": "4.1.0",
+    "@nebular/bootstrap": "4.1.0",
+    "@nebular/eva-icons": "4.1.0",
+    "@nebular/security": "4.1.0",
+    "@nebular/theme": "4.1.0",
     "@swimlane/ngx-charts": "^10.0.0",
     "angular-tree-component": "7.2.0",
     "angular2-chartjs": "0.4.1",
     "angular2-toaster": "^7.0.0",
-    "bootstrap": "4.0.0",
+    "bootstrap": "4.3.1",
     "chart.js": "2.7.1",
     "ckeditor": "4.7.3",
     "classlist.js": "1.1.20150312",
@@ -70,23 +64,24 @@
     "ng2-completer": "2.0.8",
     "ng2-smart-table": "1.3.5",
     "ngx-echarts": "^4.0.1",
+    "node-sass": "^4.12.0",
     "normalize.css": "6.0.0",
     "pace-js": "1.0.2",
     "roboto-fontface": "0.8.0",
-    "rxjs": "6.4.0",
+    "rxjs": "6.5.2",
     "rxjs-compat": "6.3.0",
     "socicon": "3.0.5",
     "tinymce": "4.5.7",
     "tslib": "^1.9.0",
     "typeface-exo": "0.0.22",
-    "web-animations-js": "2.2.5",
-    "zone.js": "^0.8.29"
+    "web-animations-js": "github:angular/web-animations-js#release_pr208",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.12.1",
-    "@angular/compiler-cli": "^7.2.11",
-    "@angular/cli": "^7.3.7",
-    "@angular/language-service": "7.2.11",
+    "@angular-devkit/build-angular": "~0.800.2",
+    "@angular/cli": "^8.0.2",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "8.0.0",
     "@compodoc/compodoc": "1.0.1",
     "@fortawesome/fontawesome-free": "^5.2.0",
     "@types/d3-color": "1.0.5",
@@ -95,7 +90,7 @@
     "@types/jasminewd2": "2.0.3",
     "@types/leaflet": "1.2.3",
     "@types/node": "6.0.90",
-    "codelyzer": "^4.5.0",
+    "codelyzer": "^5.0.1",
     "conventional-changelog-cli": "1.3.4",
     "husky": "0.13.3",
     "jasmine-core": "2.6.4",
@@ -111,8 +106,8 @@
     "rimraf": "2.6.1",
     "stylelint": "7.13.0",
     "ts-node": "3.2.2",
-    "tslint": "5.7.0",
+    "tslint": "^5.7.0",
     "tslint-language-service": "^0.9.9",
-    "typescript": "3.2.4"
+    "typescript": "3.4.5"
   }
 }
