@@ -1,179 +1,226 @@
+import { NbJSThemeOptions } from '@nebular/theme';
+
+const palette = {
+  primary: '#73a1ff',
+  success: '#5dcfe3',
+  info: '#ba7fec',
+  warning: '#ffa36b',
+  danger: '#ff6b83',
+};
+
+const theme = {
+  fontMain: 'Open Sans, sans-serif',
+  fontSecondary: 'Raleway, sans-serif',
+
+  bg: '#ffffff',
+  bg2: '#f7f9fc',
+  bg3: '#edf1f7',
+  bg4: '#e4e9f2',
+
+  border: '#ffffff',
+  border2: '#f7f9fc',
+  border3: '#edf1f7',
+  border4: '#e4e9f2',
+  border5: '#c5cee0',
+
+  fg: '#8f9bb3',
+  fgHeading: '#1a2138',
+  fgText: '#1a2138',
+  fgHighlight: palette.primary,
+  layoutBg: '#f7f9fc',
+  separator: '#edf1f7',
+
+  primary: palette.primary,
+  success: palette.success,
+  info: palette.info,
+  warning: palette.warning,
+  danger: palette.danger,
+
+  primaryLight: '#598bff',
+  successLight: '#2ce69b',
+  infoLight: '#42aaff',
+  warningLight: '#ffc94d',
+  dangerLight: '#ff708d',
+};
+
 export const CORPORATE_THEME = {
   name: 'corporate',
-  base: 'default',
   variables: {
-    temperature: [
-      '#ffa36b',
-      '#ffa36b',
-      '#ff9e7a',
-      '#ff9888',
-      '#ff8ea0',
-    ],
+    ...theme,
+
+    temperature: {
+      arcFill: [ '#ffa36b', '#ffa36b', '#ff9e7a', '#ff9888', '#ff8ea0' ],
+      arcEmpty: theme.bg2,
+      thumbBg: theme.bg2,
+      thumbBorder: '#ffa36b',
+    },
 
     solar: {
-      gradientLeft: '#ff8ea0',
-      gradientRight: '#ffa36b',
+      gradientLeft: theme.primary,
+      gradientRight: theme.primary,
       shadowColor: 'rgba(0, 0, 0, 0)',
+      secondSeriesFill: theme.bg2,
       radius: ['80%', '90%'],
     },
 
     traffic: {
-      colorBlack: '#ffffff',
-      tooltipBg: '#eef2f5',
-      tooltipBorderColor: '#eef2f5',
+      tooltipBg: theme.bg,
+      tooltipBorderColor: theme.border2,
       tooltipExtraCss: 'border-radius: 10px; padding: 4px 16px;',
-      tooltipTextColor: '#2a2a2a',
-      tooltipFontWeight: '400',
+      tooltipTextColor: theme.fgText,
+      tooltipFontWeight: 'normal',
 
-      lineBg: '#cae6f3',
+      yAxisSplitLine: 'rgba(0, 0, 0, 0)',
+
+      lineBg: theme.primary,
       lineShadowBlur: '0',
-      itemColor: '#bcc3cc',
-      itemBorderColor: '#bcc3cc',
-      itemEmphasisBorderColor: '#74a2ff',
+      itemColor: theme.border4,
+      itemBorderColor: theme.border4,
+      itemEmphasisBorderColor: theme.primaryLight,
       shadowLineDarkBg: 'rgba(0, 0, 0, 0)',
       shadowLineShadow: 'rgba(0, 0, 0, 0)',
-      gradFrom: '#ffffff',
-      gradTo: '#ffffff',
+      gradFrom: theme.bg,
+      gradTo: theme.bg,
     },
 
     electricity: {
-      tooltipBg: '#edf0f4',
-      tooltipLineColor: '#bdc4cd',
+      tooltipBg: theme.bg,
+      tooltipLineColor: theme.fgText,
       tooltipLineWidth: '0',
-      tooltipBorderColor: '#ebeef2',
+      tooltipBorderColor: theme.border2,
       tooltipExtraCss: 'border-radius: 10px; padding: 8px 24px;',
-      tooltipTextColor: '#2a2a2a',
-      tooltipFontWeight: 'bolder',
+      tooltipTextColor: theme.fgText,
+      tooltipFontWeight: 'normal',
 
-      axisLineColor: 'rgba(0, 0, 0, 0)',
-      xAxisTextColor: '#2a2a2a',
-      yAxisSplitLine: '#ebeef2',
+      axisLineColor: theme.border3,
+      xAxisTextColor: theme.fg,
+      yAxisSplitLine: theme.separator,
 
-      itemBorderColor: '#73a1ff',
+      itemBorderColor: theme.primary,
       lineStyle: 'solid',
       lineWidth: '4',
-      lineGradFrom: '#bdc4cd',
-      lineGradTo: '#c0c8d1',
+      lineGradFrom: theme.primary,
+      lineGradTo: theme.primary,
       lineShadow: 'rgba(0, 0, 0, 0)',
 
-      areaGradFrom: 'rgba(255, 255, 255, 0)',
-      areaGradTo: 'rgba(255, 255, 255, 0)',
-      shadowLineDarkBg: 'rgba(255, 255, 255, 0)',
+      areaGradFrom: 'rgba(0, 0, 0, 0)',
+      areaGradTo: 'rgba(0, 0, 0, 0)',
+      shadowLineDarkBg: 'rgba(0, 0, 0, 0)',
     },
 
     bubbleMap: {
-      titleColor: '#484848',
-      areaColor: '#dddddd',
-      areaHoverColor: '#cccccc',
-      areaBorderColor: '#ebeef2',
+      titleColor: theme.fgText,
+      areaColor: theme.bg4,
+      areaHoverColor: theme.fgHighlight,
+      areaBorderColor: theme.border5,
     },
 
     profitBarAnimationEchart: {
-      textColor: '#b2bac2',
+      textColor: theme.fgText,
 
-      firstAnimationBarColor: '#719efc',
-      secondAnimationBarColor: '#5dcfe3',
+      firstAnimationBarColor: theme.primary,
+      secondAnimationBarColor: theme.success,
 
-      splitLineStyleOpacity: '0.06',
+      splitLineStyleOpacity: '1',
       splitLineStyleWidth: '1',
-      splitLineStyleColor: '#000000',
+      splitLineStyleColor: theme.separator,
 
-      tooltipTextColor: '#2a2a2a',
-      tooltipFontWeight: '400',
+      tooltipTextColor: theme.fgText,
+      tooltipFontWeight: 'normal',
       tooltipFontSize: '16',
-      tooltipBg: '#eef2f5',
-      tooltipBorderColor: '#eef2f5',
-      tooltipBorderWidth: '3',
+      tooltipBg: theme.bg,
+      tooltipBorderColor: theme.border2,
+      tooltipBorderWidth: '1',
       tooltipExtraCss: 'border-radius: 10px; padding: 4px 16px;',
     },
 
     trafficBarEchart: {
-      gradientFrom: '#ff8ea0',
-      gradientTo: '#ffa36b',
-      shadow: 'rgba(0, 0, 0, 0)',
+      gradientFrom: theme.warningLight,
+      gradientTo: theme.warning,
+      shadow: theme.warningLight,
       shadowBlur: '0',
 
-      axisTextColor: '#b2bac2',
+      axisTextColor: theme.fgText,
       axisFontSize: '12',
 
-      tooltipBg: '#edf0f4',
-      tooltipBorderColor: '#ebeef2',
+      tooltipBg: theme.bg,
+      tooltipBorderColor: theme.border2,
       tooltipExtraCss: 'border-radius: 10px; padding: 8px 24px;',
-      tooltipTextColor: '#2a2a2a',
-      tooltipFontWeight: 'bolder',
+      tooltipTextColor: theme.fgText,
+      tooltipFontWeight: 'normal',
     },
 
     countryOrders: {
-      countryBorderColor: 'rgba(255, 255, 255, 1)',
-      countryFillColor: 'rgba(236, 242, 245, 1)',
+      countryBorderColor: theme.border4,
+      countryFillColor: theme.bg4,
       countryBorderWidth: '1',
-      hoveredCountryBorderColor: 'rgba(113, 158, 252, 1)',
-      hoveredCountryFillColor: 'rgba(199, 216, 247, 1)',
-      hoveredCountryBorderWidth: '3',
+      hoveredCountryBorderColor: theme.primary,
+      hoveredCountryFillColor: theme.primaryLight,
+      hoveredCountryBorderWidth: '1',
 
-      chartAxisLineColor: 'rgba(0, 0, 0, 0)',
-      chartAxisTextColor: '#b2bac2',
+      chartAxisLineColor: theme.border4,
+      chartAxisTextColor: theme.fg,
       chartAxisFontSize: '16',
-      chartGradientTo: 'rgba(113, 158, 252, 1)',
-      chartGradientFrom: 'rgba(113, 158, 252, 1)',
-      chartAxisSplitLine: '#ebeef2',
-      chartShadowLineColor: '#2f296b',
+      chartGradientTo: theme.primary,
+      chartGradientFrom: theme.primaryLight,
+      chartAxisSplitLine: theme.separator,
+      chartShadowLineColor: theme.primaryLight,
 
-      chartLineBottomShadowColor: 'rgba(113, 158, 252, 1)',
+      chartLineBottomShadowColor: theme.primary,
 
-      chartInnerLineColor: '#eceff4',
+      chartInnerLineColor: theme.bg2,
     },
 
     echarts: {
-      bg: '#ffffff',
-      textColor: '#484848',
-      axisLineColor: '#bbbbbb',
-      splitLineColor: '#ebeef2',
+      bg: theme.bg,
+      textColor: theme.fgText,
+      axisLineColor: theme.fgText,
+      splitLineColor: theme.separator,
       itemHoverShadowColor: 'rgba(0, 0, 0, 0.5)',
-      tooltipBackgroundColor: '#6a7985',
+      tooltipBackgroundColor: theme.primary,
       areaOpacity: '0.7',
     },
 
     chartjs: {
-      axisLineColor: '#cccccc',
-      textColor: '#484848',
+      axisLineColor: theme.separator,
+      textColor: theme.fgText,
     },
 
     orders: {
-      tooltipBg: '#ffffff',
+      tooltipBg: theme.bg,
       tooltipLineColor: 'rgba(0, 0, 0, 0)',
       tooltipLineWidth: '0',
-      tooltipBorderColor: '#ebeef2',
+      tooltipBorderColor: theme.border2,
       tooltipExtraCss: 'border-radius: 10px; padding: 8px 24px;',
-      tooltipTextColor: '#2a2a2a',
-      tooltipFontWeight: 'bolder',
+      tooltipTextColor: theme.fgText,
+      tooltipFontWeight: 'normal',
       tooltipFontSize: '20',
 
-      axisLineColor: 'rgba(161, 161 ,229, 0.3)',
+      axisLineColor: theme.border4,
       axisFontSize: '16',
-      axisTextColor: '#b2bac2',
-      yAxisSplitLine: 'rgba(161, 161 ,229, 0.2)',
+      axisTextColor: theme.fg,
+      yAxisSplitLine: theme.separator,
 
-      itemBorderColor: '#73a1ff',
+      itemBorderColor: theme.primary,
       lineStyle: 'solid',
       lineWidth: '4',
 
       // first line
-      firstAreaGradFrom: 'rgba(227, 236, 254, 0.7)',
-      firstAreaGradTo: 'rgba(227, 236, 254, 0.7)',
+      firstAreaGradFrom: theme.bg3,
+      firstAreaGradTo: theme.bg3,
       firstShadowLineDarkBg: 'rgba(0, 0, 0, 0)',
 
       // second line
-      secondLineGradFrom: 'rgba(93, 207, 227, 1)',
-      secondLineGradTo: 'rgba(93, 207, 227, 1)',
+      secondLineGradFrom: theme.primary,
+      secondLineGradTo: theme.primary,
 
       secondAreaGradFrom: 'rgba(0, 0, 0, 0)',
       secondAreaGradTo: 'rgba(0, 0, 0, 0)',
       secondShadowLineDarkBg: 'rgba(0, 0, 0, 0)',
 
       // third line
-      thirdLineGradFrom: 'rgba(113, 158, 252, 1)',
-      thirdLineGradTo: 'rgba(113, 158, 252, 1)',
+      thirdLineGradFrom: theme.success,
+      thirdLineGradTo: theme.successLight,
 
       thirdAreaGradFrom: 'rgba(0, 0, 0, 0)',
       thirdAreaGradTo: 'rgba(0, 0, 0, 0)',
@@ -181,92 +228,91 @@ export const CORPORATE_THEME = {
     },
 
     profit: {
-      bg: '#ffffff',
-      textColor: '#ffffff',
-      axisLineColor: 'rgba(161, 161 ,229, 0.3)',
-      splitLineColor: 'rgba(161, 161 ,229, 0.2)',
+      bg: theme.bg,
+      textColor: theme.fgText,
+      axisLineColor: theme.border4,
+      splitLineColor: theme.separator,
       areaOpacity: '1',
 
       axisFontSize: '16',
-      axisTextColor: '#b2bac2',
+      axisTextColor: theme.fg,
 
       // first bar
-      firstLineGradFrom: '#719efc',
-      firstLineGradTo: '#719efc',
-      firstLineShadow: 'rgba(14, 16, 48, 0.4)',
+      firstLineGradFrom: theme.bg3,
+      firstLineGradTo: theme.bg3,
+      firstLineShadow: 'rgba(0, 0, 0, 0)',
 
       // second bar
-      secondLineGradFrom: '#5dcfe3',
-      secondLineGradTo: '#5dcfe3',
-      secondLineShadow: 'rgba(14, 16, 48, 0.4)',
+      secondLineGradFrom: theme.primary,
+      secondLineGradTo: theme.primary,
+      secondLineShadow: 'rgba(0, 0, 0, 0)',
 
       // third bar
-      thirdLineGradFrom: '#e3ecfe',
-      thirdLineGradTo: '#e3ecfe',
-      thirdLineShadow: 'rgba(14, 16, 48, 0.4)',
+      thirdLineGradFrom: theme.success,
+      thirdLineGradTo: theme.success,
+      thirdLineShadow: 'rgba(0, 0, 0, 0)',
     },
 
     orderProfitLegend: {
-      firstItem: '#719efc',
-      secondItem: '#5dcfe3',
-      thirdItem: '#e3ecfe',
+      firstItem: theme.success,
+      secondItem: theme.primary,
+      thirdItem: theme.bg3,
     },
 
     visitors: {
-      tooltipBg: '#ffffff',
+      tooltipBg: theme.bg,
       tooltipLineColor: 'rgba(0, 0, 0, 0)',
-      tooltipLineWidth: '0',
-      tooltipBorderColor: '#ebeef2',
+      tooltipLineWidth: '1',
+      tooltipBorderColor: theme.border2,
       tooltipExtraCss: 'border-radius: 10px; padding: 8px 24px;',
-      tooltipTextColor: '#2a2a2a',
-      tooltipFontWeight: 'bolder',
+      tooltipTextColor: theme.fgText,
+      tooltipFontWeight: 'normal',
       tooltipFontSize: '20',
 
-      axisLineColor: 'rgba(161, 161 ,229, 0.3)',
+      axisLineColor: theme.border4,
       axisFontSize: '16',
-      axisTextColor: '#b2bac2',
-      yAxisSplitLine: 'rgba(161, 161 ,229, 0.2)',
+      axisTextColor: theme.fg,
+      yAxisSplitLine: theme.separator,
 
-      itemBorderColor: '#73a1ff',
+      itemBorderColor: theme.primary,
       lineStyle: 'dotted',
       lineWidth: '6',
-      lineGradFrom: '#73a1ff',
-      lineGradTo: '#73a1ff',
+      lineGradFrom: '#ffffff',
+      lineGradTo: '#ffffff',
       lineShadow: 'rgba(0, 0, 0, 0)',
 
-      areaGradFrom: 'rgba(146, 181, 252, 1)',
-      areaGradTo: 'rgba(146, 181, 252, 1)',
-      shadowLineDarkBg: '#a695ff',
+      areaGradFrom: theme.primary,
+      areaGradTo: theme.primaryLight,
 
       innerLineStyle: 'solid',
       innerLineWidth: '1',
 
-      innerAreaGradFrom: 'rgba(227, 236, 254, 1)',
-      innerAreaGradTo: 'rgba(227, 236, 254, 1)',
+      innerAreaGradFrom: theme.success,
+      innerAreaGradTo: theme.success,
     },
 
     visitorsLegend: {
-      firstIcon: '#e3ecfe',
-      secondIcon: '#719efc',
+      firstIcon: theme.success,
+      secondIcon: theme.primary,
     },
 
     visitorsPie: {
-      firstPieGradientLeft: '#94e2ed',
-      firstPieGradientRight: '#94e2ed',
+      firstPieGradientLeft: theme.success,
+      firstPieGradientRight: theme.success,
       firstPieShadowColor: 'rgba(0, 0, 0, 0)',
       firstPieRadius: ['65%', '90%'],
 
-      secondPieGradientLeft: '#719efc',
-      secondPieGradientRight: '#719efc',
-      secondPieShadowColor: '#b2cafb',
+      secondPieGradientLeft: theme.warning,
+      secondPieGradientRight: theme.warningLight,
+      secondPieShadowColor: 'rgba(0, 0, 0, 0)',
       secondPieRadius: ['63%', '92%'],
       shadowOffsetX: '-4',
       shadowOffsetY: '-4',
     },
 
     visitorsPieLegend: {
-      firstSection: '#719efc',
-      secondSection: '#99e5ee',
+      firstSection: theme.warning,
+      secondSection: theme.success,
     },
 
     earningPie: {
@@ -275,30 +321,30 @@ export const CORPORATE_THEME = {
 
       fontSize: '22',
 
-      firstPieGradientLeft: '#719efc',
-      firstPieGradientRight: '#719efc',
+      firstPieGradientLeft: theme.success,
+      firstPieGradientRight: theme.success,
       firstPieShadowColor: 'rgba(0, 0, 0, 0)',
 
-      secondPieGradientLeft: '#ff9f6f',
-      secondPieGradientRight: '#ff9f6f',
+      secondPieGradientLeft: theme.primary,
+      secondPieGradientRight: theme.primary,
       secondPieShadowColor: 'rgba(0, 0, 0, 0)',
 
-      thirdPieGradientLeft: '#ff5e83',
-      thirdPieGradientRight: '#ff5e83',
+      thirdPieGradientLeft: theme.warning,
+      thirdPieGradientRight: theme.warning,
       thirdPieShadowColor: 'rgba(0, 0, 0, 0)',
     },
 
     earningLine: {
-      gradFrom: '#e3ecfe',
-      gradTo: '#e3ecfe',
+      gradFrom: theme.primary,
+      gradTo: theme.primary,
 
-      tooltipTextColor: '#2a2a2a',
-      tooltipFontWeight: '400',
+      tooltipTextColor: theme.fgText,
+      tooltipFontWeight: 'normal',
       tooltipFontSize: '16',
-      tooltipBg: '#eef2f5',
-      tooltipBorderColor: '#eef2f5',
-      tooltipBorderWidth: '3',
+      tooltipBg: theme.bg,
+      tooltipBorderColor: theme.border2,
+      tooltipBorderWidth: '1',
       tooltipExtraCss: 'border-radius: 10px; padding: 4px 16px;',
     },
   },
-};
+} as NbJSThemeOptions;
