@@ -1,14 +1,22 @@
-import { Component, EventEmitter, Output } from '@angular/core';
+import { Component, EventEmitter, HostBinding, OnDestroy, OnInit, Output } from '@angular/core';
+import { Location, LocationStrategy } from '@angular/common';
+import { NbThemeService } from '@nebular/theme';
+import { map, takeUntil } from 'rxjs/operators';
+import { Subject } from 'rxjs';
 
 @Component({
   selector: 'ngx-room-selector',
   templateUrl: './room-selector.component.html',
   styleUrls: ['./room-selector.component.scss'],
 })
-export class RoomSelectorComponent {
+export class RoomSelectorComponent implements OnInit, OnDestroy {
+
+  private destroy$ = new Subject<void>();
+  private hideGrid: boolean;
+
   @Output() select: EventEmitter<number> = new EventEmitter();
 
-  selectedRoom: null;
+  selectedRoom = null;
   sortedRooms = [];
   viewBox = '-20 -20 618.88 407.99';
   isIE = !!(navigator.userAgent.match(/Trident/)
@@ -58,8 +66,33 @@ export class RoomSelectorComponent {
     ],
   };
 
-  constructor() {
+  @HostBinding('style.background')
+  get background(): 'none' | null {
+    return this.hideGrid ? 'none' : null;
+  }
+
+  constructor(
+    private location: Location,
+    private locationStrategy: LocationStrategy,
+    private themeService: NbThemeService,
+  ) {
     this.selectRoom('2');
+  }
+
+  ngOnInit() {
+    this.hideGrid = this.themeService.currentTheme === 'corporate';
+
+    this.themeService.onThemeChange()
+      .pipe(
+        map(({ name }) => name === 'corporate'),
+        takeUntil(this.destroy$),
+      )
+      .subscribe((hideGrid: boolean) => this.hideGrid = hideGrid);
+  }
+
+  ngOnDestroy() {
+    this.destroy$.next();
+    this.destroy$.complete();
   }
 
   private sortRooms() {
@@ -78,5 +111,15 @@ export class RoomSelectorComponent {
     this.select.emit(roomNumber);
     this.selectedRoom = roomNumber;
     this.sortRooms();
+  }
+
+  getUrlPath(id: string) {
+    let baseHref = this.locationStrategy.getBaseHref();
+    if (baseHref.endsWith('')) {
+      baseHref = baseHref.slice(0, -1);
+    }
+    const path = this.location.path();
+
+    return `url(${baseHref}${path}${id})`;
   }
 }
