@@ -6,13 +6,13 @@ import { Component, Input } from '@angular/core';
   template: `
     <nb-card (click)="on = !on" [ngClass]="{'off': !on}">
       <div class="icon-container">
-        <div class="icon {{ type }}">
+        <div class="icon status-{{ type }}">
           <ng-content></ng-content>
         </div>
       </div>
 
       <div class="details">
-        <div class="title">{{ title }}</div>
+        <div class="title h5">{{ title }}</div>
         <div class="status">{{ on ? 'ON' : 'OFF' }}</div>
       </div>
     </nb-card>
