@@ -5,9 +5,9 @@ import { Component, Input } from '@angular/core';
   styleUrls: ['./chart-panel-summary.component.scss'],
   template: `
     <div class="summary-container">
-      <div class="summory" *ngFor="let item of summary">
-        <div class="title">{{ item.title }}</div>
-        <div class="value">{{ item.value }}</div>
+      <div *ngFor="let item of summary">
+        <div>{{ item.title }}</div>
+        <div class="h6">{{ item.value }}</div>
       </div>
     </div>
   `,
