@@ -9,10 +9,6 @@ export class InteractiveProgressBarComponent {
 
   value = 25;
 
-  setValue(newValue) {
-    this.value = Math.min(Math.max(newValue, 0), 100);
-  }
-
   get status() {
     if (this.value <= 25) {
       return 'danger';
@@ -22,6 +18,26 @@ export class InteractiveProgressBarComponent {
       return 'info';
     } else {
       return 'success';
+    }
+  }
+
+  get canIncrease(): boolean {
+    return this.value < 100;
+  }
+
+  get canDecrease(): boolean {
+    return this.value > 0;
+  }
+
+  decreaseValue() {
+    if (this.value > 0) {
+      this.value -= 25;
+    }
+  }
+
+  increaseValue() {
+    if (this.value < 100) {
+      this.value += 25;
     }
   }
 }
