@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
+import { RouterModule, Routes } from '@angular/router';
 
 import { FormsComponent } from './forms.component';
 import { FormInputsComponent } from './form-inputs/form-inputs.component';
@@ -7,32 +7,34 @@ import { FormLayoutsComponent } from './form-layouts/form-layouts.component';
 import { DatepickerComponent } from './datepicker/datepicker.component';
 import { ButtonsComponent } from './buttons/buttons.component';
 
-const routes: Routes = [{
-  path: '',
-  component: FormsComponent,
-  children: [
-    {
-      path: 'inputs',
-      component: FormInputsComponent,
-    },
-    {
-      path: 'layouts',
-      component: FormLayoutsComponent,
-    },
-    {
-      path: 'layouts',
-      component: FormLayoutsComponent,
-    },
-    {
-      path: 'buttons',
-      component: ButtonsComponent,
-    },
-    {
-      path: 'datepicker',
-      component: DatepickerComponent,
-    },
-  ],
-}];
+const routes: Routes = [
+  {
+    path: '',
+    component: FormsComponent,
+    children: [
+      {
+        path: 'inputs',
+        component: FormInputsComponent,
+      },
+      {
+        path: 'layouts',
+        component: FormLayoutsComponent,
+      },
+      {
+        path: 'layouts',
+        component: FormLayoutsComponent,
+      },
+      {
+        path: 'buttons',
+        component: ButtonsComponent,
+      },
+      {
+        path: 'datepicker',
+        component: DatepickerComponent,
+      },
+    ],
+  },
+];
 
 @NgModule({
   imports: [
@@ -43,12 +45,5 @@ const routes: Routes = [{
   ],
 })
 export class FormsRoutingModule {
-
 }
 
-export const routedComponents = [
-  FormsComponent,
-  FormInputsComponent,
-  FormLayoutsComponent,
-  DatepickerComponent,
-];
