@@ -1,17 +1,45 @@
 import { NgModule } from '@angular/core';
+import {
+  NbActionsModule,
+  NbButtonModule,
+  NbCardModule,
+  NbCheckboxModule,
+  NbDatepickerModule, NbIconModule,
+  NbInputModule,
+  NbRadioModule,
+  NbSelectModule,
+  NbUserModule,
+} from '@nebular/theme';
 
 import { ThemeModule } from '../../@theme/theme.module';
-import { FormsRoutingModule, routedComponents } from './forms-routing.module';
-import { ButtonsModule } from './buttons/buttons.module';
+import { FormsRoutingModule } from './forms-routing.module';
+import { FormsComponent } from './forms.component';
+import { FormInputsComponent } from './form-inputs/form-inputs.component';
+import { FormLayoutsComponent } from './form-layouts/form-layouts.component';
+import { DatepickerComponent } from './datepicker/datepicker.component';
+import { ButtonsComponent } from './buttons/buttons.component';
 
 @NgModule({
   imports: [
     ThemeModule,
+    NbInputModule,
+    NbCardModule,
+    NbButtonModule,
+    NbActionsModule,
+    NbUserModule,
+    NbCheckboxModule,
+    NbRadioModule,
+    NbDatepickerModule,
     FormsRoutingModule,
-    ButtonsModule,
+    NbSelectModule,
+    NbIconModule,
   ],
   declarations: [
-    ...routedComponents,
+    FormsComponent,
+    ButtonsComponent,
+    FormInputsComponent,
+    FormLayoutsComponent,
+    DatepickerComponent,
   ],
 })
 export class FormsModule { }
