@@ -1,14 +1,20 @@
 import { NgModule } from '@angular/core';
+import { NbCardModule } from '@nebular/theme';
+
 import { ThemeModule } from '../../@theme/theme.module';
-import { MiscellaneousRoutingModule, routedComponents } from './miscellaneous-routing.module';
+import { MiscellaneousRoutingModule } from './miscellaneous-routing.module';
+import { MiscellaneousComponent } from './miscellaneous.component';
+import { NotFoundComponent } from './not-found/not-found.component';
 
 @NgModule({
   imports: [
     ThemeModule,
+    NbCardModule,
     MiscellaneousRoutingModule,
   ],
   declarations: [
-    ...routedComponents,
+    MiscellaneousComponent,
+    NotFoundComponent,
   ],
 })
 export class MiscellaneousModule { }
