@@ -1,26 +1,24 @@
 import { NgModule } from '@angular/core';
+import { NbMenuModule } from '@nebular/theme';
 
+import { ThemeModule } from '../@theme/theme.module';
 import { PagesComponent } from './pages.component';
 import { DashboardModule } from './dashboard/dashboard.module';
 import { ECommerceModule } from './e-commerce/e-commerce.module';
 import { PagesRoutingModule } from './pages-routing.module';
-import { ThemeModule } from '../@theme/theme.module';
 import { MiscellaneousModule } from './miscellaneous/miscellaneous.module';
-
-const PAGES_COMPONENTS = [
-  PagesComponent,
-];
 
 @NgModule({
   imports: [
     PagesRoutingModule,
     ThemeModule,
+    NbMenuModule,
     DashboardModule,
     ECommerceModule,
     MiscellaneousModule,
   ],
   declarations: [
-    ...PAGES_COMPONENTS,
+    PagesComponent,
   ],
 })
 export class PagesModule {
