@@ -3,14 +3,21 @@ import { Routes, RouterModule } from '@angular/router';
 
 import { TablesComponent } from './tables.component';
 import { SmartTableComponent } from './smart-table/smart-table.component';
+import { TreeGridComponent } from './tree-grid/tree-grid.component';
 
 const routes: Routes = [{
   path: '',
   component: TablesComponent,
-  children: [{
-    path: 'smart-table',
-    component: SmartTableComponent,
-  }],
+  children: [
+    {
+      path: 'smart-table',
+      component: SmartTableComponent,
+    },
+    {
+      path: 'tree-grid',
+      component: TreeGridComponent,
+    },
+  ],
 }];
 
 @NgModule({
@@ -22,4 +29,5 @@ export class TablesRoutingModule { }
 export const routedComponents = [
   TablesComponent,
   SmartTableComponent,
+  TreeGridComponent,
 ];
